+++ indico/util/passwords.py
@@ -150,7 +150,7 @@ def check_password_pwned(password, *, fast=False):
     return sha[5:] in hashes
 
 
-def validate_secure_password(context, password, *, username='', fast=False):
+def validate_secure_password(context, password, *, username='', emails=frozenset(), fast=False):
     """Check if a password is considered secure.
 
     A password is considered secure if it:
@@ -163,6 +163,7 @@ def validate_secure_password(context, password, *, username='', fast=False):
     :param context: A string indicating the context where the password is used
     :param password: The plaintext password
     :param username: The corresponding username (may be empty if not applicable)
+    :param emails: A set of email addresses (that may not be in the password)
     :param fast: Whether the check should finish quickly, even if that may
                  indicate not being able to check the password against the list
                  of pwned passwords. This should be used during interactive requests
@@ -191,6 +192,9 @@ def validate_secure_password(context, password, *, username='', fast=False):
 
     if len(username) >= 5 and len(password) <= 16 and username.lower() in password.lower():
         return _('Passwords may not contain your username.')
+
+    if any(x in password for x in emails):
+        return _('Passwords may not contain your email address.')
 
     if check_password_pwned(password, fast=fast):
         return _('This password has been seen in previous data breaches.')